refactor: coffeechat_service -> coffee_chat_service
